requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.1
fake-useragent==1.5.1
requests==2.32.2
mysql-connector-python==8.3.0
dateparser==1.2.0
tzlocal==5.2
